# openapi.yaml
openapi: 3.1.0
info:
  title: BetterReads API
  description: Book tracking and review API.
  contact:
    name: BetterReads
  license:
    name: Apache 2.0
    url: https://github.com/dahl-jar/betterreads/blob/main/LICENSE
  version: v1
servers:
- url: http://localhost:8080
  description: Production
security:
- bearer-jwt: []
...
paths:
...
  /api/v1/books/{key}/reviews/me:
    put:
      tags:
      - Reviews
      summary: Post or edit the caller's review of a book
      operationId: upsert
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertReviewRequest"
        required: true
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: The saved review
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ReviewResponse"
        "400":
          description: Rating out of range or text too long
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: No book with that key
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    delete:
      tags:
      - Reviews
      summary: Remove the caller's review of a book
      operationId: remove
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "503":
          description: Service Unavailable
        "204":
          description: Review removed or already absent
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: No book with that key
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/v1/reviews/{reviewId}/comments:
    get:
      tags:
      - Comments
      summary: List a review's comments
      operationId: listForReview
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          maximum: 100
          minimum: 1
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: A page of comments
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CommentResponse"
                  meta:
                    $ref: "#/components/schemas/ResponseMeta"
        "404":
          description: No review with that id
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security: []
    post:
      tags:
      - Comments
      summary: Post a comment on a review
      operationId: commentOnReview
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
        required: true
      responses:
        "503":
          description: Service Unavailable
        "201":
          description: The created comment
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/CommentResponse"
        "404":
          description: "No review with that id, or no parent comment with that id"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: "Empty body, reply to a reply, or parent comment on another\
            \ book or review"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/v1/books/{key}/comments:
    get:
      tags:
      - Comments
      summary: List a book's comments
      operationId: listForBook
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          maximum: 100
          minimum: 1
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: A page of comments
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CommentResponse"
                  meta:
                    $ref: "#/components/schemas/ResponseMeta"
        "404":
          description: No book with that key
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security: []
    post:
      tags:
      - Comments
      summary: Post a comment on a book
      operationId: commentOnBook
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
        required: true
      responses:
        "503":
          description: Service Unavailable
        "201":
          description: The created comment
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/CommentResponse"
        "404":
          description: "No book with that key, or no parent comment with that id"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: "Empty body, reply to a reply, or parent comment on another\
            \ book or review"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/v1/me/reviews:
    get:
      tags:
      - Reviews
      summary: List the caller's reviews
      operationId: listOwn
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          maximum: 100
          minimum: 1
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: A page of the caller's reviews
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReviewResponse"
                  meta:
                    $ref: "#/components/schemas/ResponseMeta"
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/v1/comments/{commentId}/replies:
    get:
      tags:
      - Comments
      summary: List a comment's replies
      operationId: listReplies
      parameters:
      - name: commentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          maximum: 100
          minimum: 1
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: A page of replies
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CommentResponse"
                  meta:
                    $ref: "#/components/schemas/ResponseMeta"
      security: []
...
  /api/v1/books/{key}/reviews:
    get:
      tags:
      - Reviews
      summary: List a book's reviews
      operationId: listForBook_1
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          maximum: 100
          minimum: 1
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: A page of reviews
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReviewResponse"
                  meta:
                    $ref: "#/components/schemas/ResponseMeta"
        "404":
          description: No book with that key
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security: []
...
  /api/v1/books/{key}/community-rating:
    get:
      tags:
      - Reviews
      summary: Get a book's community rating breakdown
      operationId: communityRating
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: The community rating
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/CommunityRatingResponse"
        "404":
          description: No book with that key
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security: []
...
  /api/v1/comments/{commentId}:
    delete:
      tags:
      - Comments
      summary: Remove the caller's comment
      operationId: remove_2
      parameters:
      - name: commentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "503":
          description: Service Unavailable
        "204":
          description: Comment removed or already absent
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: The comment belongs to another user
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
components:
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
...
    UpsertReviewRequest:
      type: object
      properties:
        rating:
          type: integer
          format: int32
          maximum: 5
          minimum: 1
        title:
          type: string
          maxLength: 255
          minLength: 0
        body:
          type: string
          maxLength: 5000
          minLength: 0
      required:
      - rating
    ReviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bookKey:
          type: string
        rating:
          type: integer
          format: int32
        title:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date
    CreateCommentRequest:
      type: object
      properties:
        body:
          type: string
          maxLength: 5000
          minLength: 0
        parentCommentId:
          type: integer
          format: int64
      required:
      - body
    CommentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        body:
          type: string
        author:
          type: string
        createdAt:
          type: string
          format: date
        replyCount:
          type: integer
          format: int64
...
    CommunityRatingResponse:
      type: object
      properties:
        average:
          type: number
        count:
          type: integer
          format: int64
        distribution:
          type: array
          items:
            $ref: "#/components/schemas/StarCount"
    StarCount:
      type: object
      properties:
        star:
          type: integer
          format: int32
        count:
          type: integer
          format: int64
...
    ResponseMeta:
      properties:
        total:
          type: integer
          format: int64
          example: 142
        offset:
          type: integer
          format: int32
          example: 0
        limit:
          type: integer
          format: int32
          example: 20
  securitySchemes:
    bearer-jwt:
      type: http
      description: Access JWT from POST /api/v1/auth/login.
      scheme: bearer
      bearerFormat: JWT